move scrollToElement from textAnalysisUtils.jsx to new domUtils.js
