+++ src/components/NavItems/Assistant/utils/domUtils.js
@@ -1,0 +1,8 @@
+export function scrollToElement(id, padding = 0) {
+  const element = document.getElementById(id);
+  if (element) {
+    const targetPosition =
+      element.getBoundingClientRect().top + window.scrollY - padding;
+    window.scrollTo({ top: targetPosition, behavior: "smooth" });
+  }
+}
