+++ src/components/NavItems/Assistant/utils/textAnalysisUtils.jsx
@@ -273,17 +273,3 @@ export function createGaugeChart(
     </>
   );
 }
-
-/**
- * Function to take user direct to a specific element on the page
- * @param id
- * @param padding
- */
-export function scrollToElement(id, padding = 0) {
-  const element = document.getElementById(id);
-  if (element) {
-    const targetPosition =
-      element.getBoundingClientRect().top + window.scrollY - padding;
-    window.scrollTo({ top: targetPosition, behavior: "smooth" });
-  }
-}
